MAINT: clean project structure.
