+++ python/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(do/sara)
