+++ python/do/sara/CMakeLists.txt
@@ -1,3 +1,4 @@
+# ============================================================================ #
 # Find the needed third-party libraries.
 find_package(NumPy 1.7.1 REQUIRED)
 find_package(PythonInterp REQUIRED)
@@ -11,17 +12,18 @@ endif ()
 find_package(Boost COMPONENTS ${SARA_BOOST_PYTHON} REQUIRED)
 
 
+# ============================================================================ #
 # Compile the library.
 include_directories(
   ${Boost_INCLUDE_DIRS}
   ${NUMPY_INCLUDE_DIRS}
   ${PYTHON_INCLUDE_DIRS})
 
-file(GLOB sara_SRC_FILES FILES *.cpp)
+file(GLOB pysara_SRC_FILES FILES *.cpp)
 
-python_add_module(sara ${sara_SRC_FILES})
+python_add_module(pysara ${pysara_SRC_FILES})
 
-target_link_libraries(sara
+target_link_libraries(pysara
   ${Boost_LIBRARIES}
   ${PYTHON_LIBRARIES}
   DO_Sara_DisjointSets
@@ -30,17 +32,24 @@ target_link_libraries(sara
   DO_Sara_VideoIO)
 
 
+# ============================================================================ #
 # Install the library in the currently active python environment.
-execute_process(COMMAND python -c
+execute_process(COMMAND ${PYTHON_EXECUTABLE} -c
   "from distutils.sysconfig import get_python_lib; print get_python_lib()"
   OUTPUT_VARIABLE PYTHON_SITE_PACKAGES_DIR
   OUTPUT_STRIP_TRAILING_WHITESPACE)
 
-set(sara_INSTALL_DIR ${PYTHON_SITE_PACKAGES_DIR}/do)
+message("Installing sara Python bindings to ${PYTHON_SITE_PACKAGES_DIR}")
 
-message("Installing sara Python bindings to ${sara_INSTALL_DIR}")
+install(TARGETS pysara
+        RUNTIME DESTINATION ${PYTHON_SITE_PACKAGES_DIR}
+        ARCHIVE DESTINATION ${PYTHON_SITE_PACKAGES_DIR}
+        LIBRARY DESTINATION ${PYTHON_SITE_PACKAGES_DIR})
 
-install(TARGETS sara
-        RUNTIME DESTINATION ${sara_INSTALL_DIR}
-        ARCHIVE DESTINATION ${sara_INSTALL_DIR}
-        LIBRARY DESTINATION ${sara_INSTALL_DIR})
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/../
+        DESTINATION ${PYTHON_SITE_PACKAGES_DIR}/do)
+
+
+if (SARA_BUILD_TESTS AND SARA_BUILD_PYTHON_BINDINGS)
+  add_subdirectory(test)
+endif ()
